return to previous directory when you exit kr
I tried using pushd and popd, but this fails when the user exits
with a ctrl-c.  I think this happens becuase the ctrl-c ends the
bash function as well as the kernel, so we never get to the popd.
It did work when the user exits cleanly (with an "exit" command),
but I prefer to exit with a ctrl-c.

By using a subprocess, when we ctrl-c the subprocess we return to
the calling shell's original directory, which is what we want.